Fix case-sensitivity issue with views and column masks
The table column reference was registered incorectly with the original
case taken from the view definition. It then failed to match the column
schema returned from `SystemAccessControl` and the mask was not applied.

Additionally, the registered table column references are also used to
check access to the columns; those were also using the original case
from the view definition and access control was not recognizing them.

Fixes #24054.